WEBAPP-104: Move i18n stuff to i18n package
