+++ src/modules/i18n/reducers/index.js
